requirements.txt
Flask==2.2.3
nltk==3.8.1
scikit-learn==1.2.2
mysql-connector-python==8.0.30
